third_party/blink/web_tests/virtual/allow-rtc-encoded-video-frame-set-metadata-all-fields/fast/peerconnection/RTCEncodedVideoFrame-set-metadata-expected.txt

This is a testharness.js-based test.
FAIL setMetadata() does not allow changes to fields without a flag assert_throws_dom: For field "decodeTargetIndications", function "() => clone.setMetadata(metadata)" did not throw
PASS setMetadata() allows changes when AllowRTCEncodedVideoFrameSetMetadataAllFields is enabled
PASS setMetadata() allows metadata when AllowRTCEncodedVideoFrameSetMetadataAllFields is enabled with no generic dependency descriptor - ie no frameId
Harness: the test ran to completion.
